pyproject.toml
[project]
name = "work-daily"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi[standard]>=0.115.12",
    "gitpython>=3.1.44",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
]
